requirements.txt
streamlit==1.32.0
moviepy==1.0.3
google-generativeai==0.3.2
python-dotenv==1.0.1
SpeechRecognition==3.10.0
pydub==0.25.1 